Restructure Scheduler page into components
