+++ src/components/TimeTable/index.tsx
@@ -1,0 +1,104 @@
+import { Box, Typography } from "@mui/material";
+import { useRef } from "react";
+import { atom, selector, useRecoilState, useRecoilValue } from "recoil";
+import Calendar from "@toast-ui/react-calendar";
+import { Events, EventsService, Timezones } from "../../services/EventsService";
+import type { EventObject } from "@toast-ui/calendar";
+
+const eventsState = atom<Events>({ key: "events" });
+const timezonesState = atom<Timezones>({ key: "timezones" });
+const intersectionsState = selector({
+    key: "intersections",
+    get: ({ get }) => {
+        const events = get(eventsState);
+        const timezones = get(timezonesState);
+        const utcEvents = EventsService.convertEventsToUTC(events, timezones);
+
+        return EventsService.mergeOverlappingIntersections(
+            EventsService.getIntersectionOfLists(
+                Object.values(utcEvents).filter((e) => e.length > 0)
+            )
+        );
+    },
+});
+
+function useEvents(id: number = 0) {
+    const intersections = useRecoilValue(intersectionsState).map((s) => ({
+        ...s,
+        calendarId: "intersection",
+    }));
+
+    const timezones = useRecoilValue(timezonesState);
+    const intersectionsInTimezone = EventsService.convertEventObjectsFromUTC(
+        intersections,
+        timezones[id]
+    );
+
+    const [globalEvents, setEvents] = useRecoilState(eventsState);
+    const selectedEvents = (globalEvents[id] || []).map((e: EventObject) => ({
+        ...e,
+        calendarId: "selection",
+    }));
+    const events = [...selectedEvents, ...intersectionsInTimezone];
+
+    function addNewEvent(event: EventObject) {
+        setEvents((allEvents) => ({
+            ...allEvents,
+            [id]: [...(allEvents[id] || []), { start: event.start, end: event.end }],
+        }));
+    }
+
+    return {
+        events,
+        addNewEvent,
+    };
+}
+
+export default function Timetable({ id }: { id: number }) {
+    const timezones = useRecoilValue(timezonesState);
+    const { events, addNewEvent } = useEvents(id);
+    const calendarRef = useRef(null);
+
+    if (!timezones[id]) {
+        return (
+            <Box flexGrow="2">
+                <Typography variant="h6">Select a timezone first.</Typography>
+            </Box>
+        );
+    }
+
+    return (
+        <Calendar
+            ref={calendarRef}
+            calendars={[
+                {
+                    id: "selection",
+                    name: "Selection",
+                    backgroundColor: "#00a9ff",
+                    borderColor: "#00a9ff",
+                },
+                {
+                    id: "intersection",
+                    name: "Intersection",
+                    backgroundColor: "#23CE6B",
+                    borderColor: "#23CE6B",
+                },
+            ]}
+            usageStatistics={false}
+            view="day"
+            week={{
+                showTimezoneCollapseButton: true,
+                timezonesCollapsed: true,
+                eventView: ["time"],
+                taskView: false,
+            }}
+            onSelectDateTime={(event) => {
+                addNewEvent(event);
+
+                const calendarInstance = (calendarRef.current as any).getInstance();
+                calendarInstance.clearGridSelections();
+            }}
+            events={events}
+        />
+    );
+}
