+++ src/index.tsx
@@ -4,6 +4,7 @@ import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import './index.css';
+import "@toast-ui/calendar/dist/toastui-calendar.min.css";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
